Fix collision event issue (partially)
